+++ src/metadata/exiv2_handler.h
@@ -43,6 +43,10 @@ class Exiv2Handler : public MediaMetadataHandler {
 public:
     explicit Exiv2Handler(const std::shared_ptr<Context>& context);
 
+    bool isSupported(const std::string& contentType,
+        bool isOggTheora,
+        const std::string& mimeType,
+        ObjectType mediaType) override;
     bool fillMetadata(const std::shared_ptr<CdsObject>& obj) override;
     std::unique_ptr<IOHandler> serveContent(
         const std::shared_ptr<CdsObject>& obj,