+++ src/metadata/libexif_handler.h
@@ -65,6 +65,10 @@ public:
     explicit LibExifHandler(const std::shared_ptr<Context>& context);
     ~LibExifHandler();
 
+    bool isSupported(const std::string& contentType,
+        bool isOggTheora,
+        const std::string& mimeType,
+        ObjectType mediaType) override;
     bool fillMetadata(const std::shared_ptr<CdsObject>& obj) override;
     std::unique_ptr<IOHandler> serveContent(
         const std::shared_ptr<CdsObject>& obj,